requirements.txt
Adafruit-BBIO==1.2.0
cryptography==1.7.1
dad==1.0.0
docutils==0.16
idna==2.2
keyring==10.1
keyrings.alt==1.3
lockfile==0.12.2
numpy==1.12.1
overrides==3.1.0
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pylibiio==0.21
python-daemon==2.3.0
python-statemachine==0.8.0
pyxdg==0.25
SecretStorage==2.3.1
six==1.10.0
speedtest-cli==2.1.2
